refactored filter transform
