+++ transforms/universal/filter/Dockerfile.python
@@ -2,9 +2,6 @@ FROM docker.io/python:3.10.14-slim-bullseye
 
 RUN pip install --upgrade --no-cache-dir pip 
 
-# install pytest
-RUN pip install --no-cache-dir pytest
-
 # Create a user and use it to run the transform
 RUN useradd -ms /bin/bash dpk
 USER dpk
@@ -16,22 +13,10 @@ ARG DPK_WHEEL_FILE_NAME
 COPY --chown=dpk:root data-processing-dist data-processing-dist
 RUN  pip install data-processing-dist/${DPK_WHEEL_FILE_NAME}
 
-# END OF STEPS destined for a data-prep-kit base image 
-
-COPY --chown=dpk:root src/ src/
-COPY --chown=dpk:root pyproject.toml pyproject.toml 
+COPY --chown=dpk:root dpk_filter/ dpk_filter/
 COPY --chown=dpk:root requirements.txt requirements.txt
-RUN pip install --no-cache-dir -e .
+RUN pip install --no-cache-dir -r requirements.txt
 
-# copy the main() entry point to the image 
-COPY ./src/filter_transform_python.py .
-
-# copy some of the samples in
-COPY ./src/filter_local.py local/
-
-# copy test
-COPY test/ test/
-COPY test-data/ test-data/
 
 # Set environment
 ENV PYTHONPATH /home/dpk
