+++ transforms/universal/filter/Dockerfile.ray
@@ -1,4 +1,5 @@
 ARG BASE_IMAGE=docker.io/rayproject/ray:2.24.0-py310
+
 FROM ${BASE_IMAGE}
 
 RUN pip install --upgrade --no-cache-dir pip 
@@ -13,25 +14,9 @@ COPY --chown=ray:users data-processing-dist data-processing-dist
 RUN  pip install data-processing-dist/${DPK_WHEEL_FILE_NAME}[ray]
 
 ## Copy the python version of the tansform
-COPY --chown=ray:users python-transform/ python-transform 
-RUN cd python-transform && pip install --no-cache-dir -e .
-
-#COPY requirements.txt requirements.txt
-#RUN pip install --no-cache-dir -r  requirements.txt
-
-COPY --chown=ray:users src/ src/
-COPY --chown=ray:users pyproject.toml pyproject.toml 
-RUN pip install --no-cache-dir -e .
-
-# copy the main() entry point to the image 
-COPY ./src/filter_transform_ray.py .
-
-# copy some of the samples in
-COPY src/filter_local_ray.py local/
-
-# copy test
-COPY test/ test/
-COPY test-data/ test-data/
+COPY --chown=ray:users dpk_filter/ dpk_filter/
+COPY --chown=ray:users requirements.txt requirements.txt
+RUN pip install -r requirements.txt
 
 # Grant non-root users the necessary permissions to the ray directory
 RUN chmod 755 /home/ray
