+++ transforms/universal/filter/dpk_filter/local_python.py
@@ -13,14 +13,14 @@
 import os
 import sys
 
+from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.utils import ParamsUtils
-from data_processing_spark.runtime.spark import SparkTransformLauncher
-from filter_transform import (
+from transforms.universal.filter.dpk_filter.transform import (
     filter_columns_to_drop_cli_param,
     filter_criteria_cli_param,
     filter_logical_operator_cli_param,
 )
-from filter_transform_spark import FilterSparkTransformConfiguration
+from dpk_filter.transform_python import FilterPythonTransformConfiguration
 
 
 # create parameters
@@ -55,6 +55,6 @@ if __name__ == "__main__":
     # Set the simulated command line args
     sys.argv = ParamsUtils.dict_to_req(d=params | filter_params)
     # create launcher
-    launcher = SparkTransformLauncher(runtime_config=FilterSparkTransformConfiguration())
+    launcher = PythonTransformLauncher(FilterPythonTransformConfiguration())
     # Launch the ray actor(s) to process the input
     launcher.launch()
