+++ transforms/universal/filter/dpk_filter/ray/local.py
@@ -13,24 +13,24 @@
 import os
 import sys
 
-from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.utils import ParamsUtils
-from filter_transform import (
+from data_processing_ray.runtime.ray import RayTransformLauncher
+from dpk_filter.transform import (
     filter_columns_to_drop_cli_param,
     filter_criteria_cli_param,
     filter_logical_operator_cli_param,
 )
-from filter_transform_python import FilterPythonTransformConfiguration
+from dpk_filter.ray.transform import FilterRayTransformConfiguration
 
 
 # create parameters
-input_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "test-data", "input"))
-output_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "output"))
+input_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "../test-data/input"))
+output_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "../output"))
 local_conf = {
     "input_folder": input_folder,
     "output_folder": output_folder,
 }
-code_location = {"github": "github", "commit_hash": "12345", "path": "path"}
+
 filter_criteria = [
     "docq_total_words > 100 AND docq_total_words < 200",
     "ibmkenlm_docq_perplex_score < 230",
@@ -43,18 +43,29 @@ filter_params = {
     filter_columns_to_drop_cli_param: filter_columns_to_drop,
     filter_logical_operator_cli_param: filter_logical_operator,
 }
-params = {
+
+worker_options = {"num_cpus": 0.8}
+code_location = {"github": "github", "commit_hash": "12345", "path": "path"}
+launcher_params = {
+    # where to run
+    "run_locally": True,
     # Data access. Only required parameters are specified
     "data_local_config": ParamsUtils.convert_to_ast(local_conf),
-    # execution info
+    # orchestrator
+    "runtime_worker_options": ParamsUtils.convert_to_ast(worker_options),
+    "runtime_num_workers": 3,
     "runtime_pipeline_id": "pipeline_id",
     "runtime_job_id": "job_id",
+    "runtime_creation_delay": 0,
     "runtime_code_location": ParamsUtils.convert_to_ast(code_location),
 }
+
+# launch
 if __name__ == "__main__":
-    # Set the simulated command line args
-    sys.argv = ParamsUtils.dict_to_req(d=params | filter_params)
-    # create launcher
-    launcher = PythonTransformLauncher(FilterPythonTransformConfiguration())
+    # Run the transform inside Ray
+    # Create the CLI args as will be parsed by the launcher
+    sys.argv = ParamsUtils.dict_to_req(launcher_params | filter_params)
+    # Create the longer to launch with the blocklist transform.
+    launcher = RayTransformLauncher(FilterRayTransformConfiguration())
     # Launch the ray actor(s) to process the input
     launcher.launch()
