+++ transforms/universal/filter/dpk_filter/spark/local.py
@@ -10,29 +10,27 @@
 # limitations under the License.
 ################################################################################
 
+import os
 import sys
 
 from data_processing.utils import ParamsUtils
-from data_processing_ray.runtime.ray import RayTransformLauncher
-from filter_transform import (
+from data_processing_spark.runtime.spark import SparkTransformLauncher
+from dpk_filter.transform import (
     filter_columns_to_drop_cli_param,
     filter_criteria_cli_param,
     filter_logical_operator_cli_param,
 )
-from filter_transform_ray import FilterRayTransformConfiguration
+from dpk_filter.spark.transform import FilterSparkTransformConfiguration
 
 
 # create parameters
-s3_cred = {
-    "access_key": "localminioaccesskey",
-    "secret_key": "localminiosecretkey",
-    "url": "http://localhost:9000",
+input_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "test-data", "input"))
+output_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "output"))
+local_conf = {
+    "input_folder": input_folder,
+    "output_folder": output_folder,
 }
-s3_conf = {
-    "input_folder": "test/filter/input",
-    "output_folder": "test/filter/output",
-}
-
+code_location = {"github": "github", "commit_hash": "12345", "path": "path"}
 filter_criteria = [
     "docq_total_words > 100 AND docq_total_words < 200",
     "ibmkenlm_docq_perplex_score < 230",
@@ -45,30 +43,18 @@ filter_params = {
     filter_columns_to_drop_cli_param: filter_columns_to_drop,
     filter_logical_operator_cli_param: filter_logical_operator,
 }
-
-worker_options = {"num_cpus": 0.8}
-code_location = {"github": "github", "commit_hash": "12345", "path": "path"}
-launcher_params = {
-    # where to run
-    "run_locally": True,
+params = {
     # Data access. Only required parameters are specified
-    "data_s3_cred": ParamsUtils.convert_to_ast(s3_cred),
-    "data_s3_config": ParamsUtils.convert_to_ast(s3_conf),
-    # orchestrator
-    "runtime_worker_options": ParamsUtils.convert_to_ast(worker_options),
-    "runtime_num_workers": 3,
+    "data_local_config": ParamsUtils.convert_to_ast(local_conf),
+    # execution info
     "runtime_pipeline_id": "pipeline_id",
     "runtime_job_id": "job_id",
-    "runtime_creation_delay": 0,
     "runtime_code_location": ParamsUtils.convert_to_ast(code_location),
 }
-
-# launch
 if __name__ == "__main__":
-    # Run the transform inside Ray
-    # Create the CLI args as will be parsed by the launcher
-    sys.argv = ParamsUtils.dict_to_req(launcher_params | filter_params)
-    # Create the longer to launch with the blocklist transform.
-    launcher = RayTransformLauncher(FilterRayTransformConfiguration())
+    # Set the simulated command line args
+    sys.argv = ParamsUtils.dict_to_req(d=params | filter_params)
+    # create launcher
+    launcher = SparkTransformLauncher(runtime_config=FilterSparkTransformConfiguration())
     # Launch the ray actor(s) to process the input
     launcher.launch()
