--- transforms/universal/filter/python/.dockerignore
@@ -1,1 +1,0 @@
-venv/
