--- transforms/universal/filter/python/requirements.txt
@@ -1,3 +1,0 @@
-
-data-prep-toolkit>=0.2.3
-duckdb>=0.10.1
