+++ transforms/universal/filter/requirements.txt
@@ -1,0 +1,2 @@
+
+duckdb>=0.10.1
