--- transforms/universal/filter/spark/.dockerignore
@@ -1,1 +1,0 @@
-venv/
