--- transforms/universal/filter/spark/pyproject.toml
@@ -1,40 +1,0 @@
-[project]
-name = "dpk_filter_transform_spark"
-version = "0.2.4.dev0"
-requires-python = ">=3.10,<3.13"
-description = "Filter Spark Transform"
-license = {text = "Apache-2.0"}
-readme = {file = "README.md", content-type = "text/markdown"}
-authors = [
-    { name = "Boris Lublinsky", email = "blublinsk@ibm.com" },
-]
-dependencies = [
-    "data-prep-toolkit[spark]>=0.2.4.dev0",
-]
-
-[project.optional-dependencies]
-dev = [
-    "twine",
-    "pytest>=7.3.2",
-    "pytest-dotenv>=0.5.2",
-    "pytest-env>=1.0.0",
-    "pre-commit>=3.3.2",
-    "pytest-cov>=4.1.0",
-    "pytest-mock>=3.10.0",
-    "moto==5.0.5",
-    "markupsafe==2.0.1",
-]
-
-[options]
-package_dir = ["src","test"]
-
-[options.packages.find]
-where = ["src/"]
-
-[tool.pytest.ini_options]
-# Currently we use low coverage since we have to run tests separately (see makefile)
-#addopts = "--cov --cov-report term-missing --cov-fail-under 25"
-markers = ["unit: unit tests", "integration: integration tests"]
-
-[tool.coverage.run]
-include = ["src/*"]
