+++ transforms/universal/filter/test/test_filter_spark.py
@@ -9,16 +9,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 ################################################################################
-
 import os
 
 from data_processing.runtime import AbstractTransformLauncher
-from data_processing_ray.runtime.ray import RayTransformLauncher
-from filter_test_support import AbstractPythonFilterTransformTest
-from filter_transform_ray import FilterRayTransformConfiguration
+from data_processing_spark.runtime.spark import SparkTransformLauncher
+from dpk_filter.test_support import AbstractPythonFilterTransformTest
+from dpk_filter.spark.transform import FilterSparkTransformConfiguration
 
 
-class TestPythonFilterTransform(AbstractPythonFilterTransformTest):
+class TestSparkFilterTransform1(AbstractPythonFilterTransformTest):
     """
     Extends the Python super-class to redefine the launcher as a RayTransformLauncher.
     The name of this class MUST begin with the word Test so that pytest recognizes it as a test class.
@@ -29,4 +28,4 @@ class TestPythonFilterTransform(AbstractPythonFilterTransformTest):
         return dir
 
     def _get_launcher(self) -> (AbstractTransformLauncher, dict):
-        return (RayTransformLauncher(FilterRayTransformConfiguration()), {"run_locally": True})
+        return (SparkTransformLauncher(FilterSparkTransformConfiguration()), {})
