--- transforms/universal/filter/transform.config
@@ -1,20 +1,0 @@
-#
-# This is intended to be included across the Makefiles provided within
-# a given transform's directory tree,  so must use compatible syntax.
-#
-################################################################################
-# This defines the name of the transform and is used to match against
-# expected files and is used to define the transform's image name. 
-TRANSFORM_NAME=filter
-
-################################################################################
-# This defines the transforms' version number as would be used
-# when publishing the wheel.  In general, only the micro version
-# number should be advanced relative to the DPK_VERSION. 
-#
-# If you change the versions numbers, be sure to run "make set-versions" to 
-# update version numbers across the transform (e.g., pyproject.toml).
-FILTER_PYTHON_VERSION=$(DPK_VERSION)
-FILTER_RAY_VERSION=$(FILTER_PYTHON_VERSION)
-FILTER_SPARK_VERSION=$(FILTER_PYTHON_VERSION)
-
